requirements.txt
aiofiles==24.1.0 ; python_version >= "3.12" and python_version < "4.0"
aiohttp==3.9.5 ; python_version >= "3.12" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.4.0 ; python_version >= "3.12" and python_version < "4.0"
attrs==23.2.0 ; python_version >= "3.12" and python_version < "4.0"
bidict==0.23.1 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.7.4 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and (platform_system == "Windows" or sys_platform == "win32")
dnspython==2.6.1 ; python_version >= "3.12" and python_version < "4.0"
docutils==0.19 ; python_version >= "3.12" and python_version < "4.0"
fastapi==0.109.2 ; python_version >= "3.12" and python_version < "4.0"
frozenlist==1.4.1 ; python_version >= "3.12" and python_version < "4.0"
ghp-import==2.1.0 ; python_version >= "3.12" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
httpcore==1.0.5 ; python_version >= "3.12" and python_version < "4.0"
httptools==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.27.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.7 ; python_version >= "3.12" and python_version < "4.0"
ifaddr==0.2.0 ; python_version >= "3.12" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
jinja2==3.1.4 ; python_version >= "3.12" and python_version < "4.0"
markdown2==2.5.0 ; python_version >= "3.12" and python_version < "4"
markdown==3.6 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.12" and python_version < "4.0"
mergedeep==1.3.4 ; python_version >= "3.12" and python_version < "4.0"
mkdocs-get-deps==0.2.0 ; python_version >= "3.12" and python_version < "4.0"
mkdocs==1.6.0 ; python_version >= "3.12" and python_version < "4.0"
multidict==6.0.5 ; python_version >= "3.12" and python_version < "4.0"
nicegui==1.4.29 ; python_version >= "3.12" and python_version < "4.0"
orjson==3.10.6 ; python_version >= "3.12" and python_version < "4.0" and (platform_machine != "i386" and platform_machine != "i686")
packaging==24.1 ; python_version >= "3.12" and python_version < "4.0"
pathspec==0.12.1 ; python_version >= "3.12" and python_version < "4.0"
platformdirs==4.2.2 ; python_version >= "3.12" and python_version < "4.0"
pscript==0.7.7 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.20.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic-settings==2.3.4 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.8.2 ; python_version >= "3.12" and python_version < "4.0"
pygments==2.18.0 ; python_version >= "3.12" and python_version < "4.0"
pymongo==4.8.0 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
python-engineio==4.9.1 ; python_version >= "3.12" and python_version < "4.0"
python-multipart==0.0.9 ; python_version >= "3.12" and python_version < "4.0"
python-socketio[asyncio-client]==5.11.3 ; python_version >= "3.12" and python_version < "4.0"
pyyaml-env-tag==0.1 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
simple-websocket==1.0.0 ; python_version >= "3.12" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.36.3 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.2.2 ; python_version >= "3.12" and python_version < "4.0"
uvicorn[standard]==0.30.3 ; python_version >= "3.12" and python_version < "4.0"
uvloop==0.19.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.12" and python_version < "4.0"
vbuild==0.8.2 ; python_version >= "3.12" and python_version < "4.0"
watchdog==4.0.1 ; python_version >= "3.12" and python_version < "4.0"
watchfiles==0.22.0 ; python_version >= "3.12" and python_version < "4.0"
websockets==12.0 ; python_version >= "3.12" and python_version < "4.0"
wsproto==1.2.0 ; python_version >= "3.12" and python_version < "4.0"
yarl==1.9.4 ; python_version >= "3.12" and python_version < "4.0"
